Adopt case-utils 0.11.0
This patch has no functional impact on the operations of this GitHub
Action, as `case_validate` was not modified between `case-utils` 0.10.0
and 0.11.0.  For adopters of this Action who also import `case-utils`
for Python reasons, the Action will be known to be running the same
`case-utils` version as among their package dependencies.

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>